restructuring and clean-up
