+++ neurobooth_os/tasks/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
-from neurobooth_os.tasks.task import Task, Task_Eyetracker
+from neurobooth_os.tasks.task import Task
+from neurobooth_os.tasks.task_eyetracker import Task_Eyetracker
