+++ neurobooth_os/tasks/sdmt.py
@@ -8,7 +8,8 @@ from psychopy.core import CountdownTimer
 from psychopy.visual import TextStim, ImageStim
 from psychopy.visual.rect import Rect
 
-from neurobooth_os.tasks.task import Eyelink_HostPC, TaskAborted, EyelinkColor
+from neurobooth_os.tasks.task import TaskAborted
+from neurobooth_os.tasks.task_eyetracker import Eyelink_HostPC, EyelinkColor
 from neurobooth_os.iout.stim_param_reader import get_cfg_path
