+++ neurobooth_os/tasks/task.py
@@ -2,35 +2,29 @@
 """
  A task is an operation or sequence of steps performed presented to a subject via Psychopy.
 """
-
 from __future__ import absolute_import, division
 
-from typing import List, Union, Optional
-from enum import Enum
+import logging
+import os.path as op
+import time
+
+import neurobooth_os
+import neurobooth_os.config as cfg
+import neurobooth_os.iout.metadator as meta
+
+from typing import List, Optional
 
 from psychopy import logging as psychopy_logging
 
 from neurobooth_os.msg.messages import StatusMessage, Request
+from neurobooth_os.tasks import utils
+from neurobooth_os.log_manager import APP_LOG_NAME
+
+from datetime import datetime
+from psychopy import visual, event
+
 
 psychopy_logging.console.setLevel(psychopy_logging.CRITICAL)
-
-import logging
-import os
-import os.path as op
-from datetime import datetime
-import time
-import pylink
-
-from psychopy import visual, monitors, sound, event
-
-import neurobooth_os
-from neurobooth_os.tasks import utils
-from neurobooth_os.tasks.smooth_pursuit.utils import deg2pix
-from neurobooth_os.log_manager import APP_LOG_NAME
-import neurobooth_os.config as cfg
-import neurobooth_os.iout.metadator as meta
-
-from pylsl import local_clock
 
 
 class TaskAborted(Exception):
@@ -56,7 +50,6 @@ class Task:
             text_continue=utils.text_continue,
             text_practice_screen=utils.text_practice_screen,
             text_task=utils.text_task,
-            text_end=utils.text_end,
             countdown=None,
             task_repeatable_by_subject: bool = True,
             **kwargs,
@@ -106,7 +99,6 @@ class Task:
             self.win_temp = False
 
         self.instruction_file = instruction_file
-        # self.load_instruction_video()
 
         # Create mouse and set not visible
         self.Mouse = event.Mouse(visible=False, win=self.win)
@@ -126,7 +118,7 @@ class Task:
         self.task_screen = utils.create_text_screen(self.win, text_task)
         self.end_screen = utils.get_end_screen(self.win)
 
-    def load_instruction_video(self):
+    def _load_instruction_video(self):
         """
         Loads instruction video if not previously loaded or if previously loaded but currently stopped.
         Returns
@@ -178,9 +170,9 @@ class Task:
         if self.with_lsl:
             self.marker.push_sample([f"{msg}_{time.time()}"])
         if add_event:
-            self.add_event(msg)
+            self._add_event(msg)
 
-    def add_event(self, event_name):
+    def _add_event(self, event_name):
         self.events.append(f"{event_name}:{datetime.now().strftime('%H:%M:%S')}")
 
     def show_text(
@@ -215,7 +207,7 @@ class Task:
             if self.repeat_advance():
                 func(**func_kwargs)
 
-    def show_video(self, video, msg, stop=False):
+    def _show_video(self, video, msg, stop=False):
         self.send_marker(f"{msg}_start", True)
         if video is not None:
             utils.play_video(self.win, video, stop=stop)
@@ -229,8 +221,8 @@ class Task:
         utils.play_tone()
 
     def present_instructions(self, prompt=True):
-        self.load_instruction_video()
-        self.show_video(video=self.instruction_video, msg="Intructions")
+        self._load_instruction_video()
+        self._show_video(video=self.instruction_video, msg="Intructions")
         if prompt:
             self.show_text(
                 screen=self.press_inst_screen,
@@ -288,319 +280,3 @@ class Task:
                 req = Request(source="STM", destination="CTR", body=msg)
                 meta.post_message(req, db_conn)
             raise TaskAborted()
-
-
-class Task_countdown(Task):
-    """
-        Task whose instances are experiments that run for a fixed amount of time (counting down in seconds).
-        It's used by speech tasks like 'La-La-La'
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    def present_task(self, prompt, duration, **kwargs):
-        self.countdown_to_stimulus()
-
-        self.send_marker(self.marker_task_start, True)
-        utils.present(self.win, self.task_screen, waitKeys=False)
-
-        duration += 2  # No idea why, but the original code was like this...
-        try:  # Keep presenting the screen until the task is over or the task is aborted.
-            start_time = local_clock()
-            while (local_clock() - start_time) < duration:
-                self.check_if_aborted()
-        except TaskAborted:
-            self.logger.info('Task aborted.')
-
-        self.win.flip()
-        self.send_marker(self.marker_task_end, True)
-
-        if prompt:
-            func_kwargs = locals()
-            del func_kwargs["self"]
-            self.show_text(
-                screen=self.press_task_screen,
-                msg="Task-continue-repeat",
-                func=self.present_task,
-                func_kwargs=func_kwargs,
-                waitKeys=False,
-            )
-
-
-class Task_pause(Task):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.screen = None
-
-    def run(self, slide_image="end_slide_01_23_25.jpg", wait_key="return", **kwarg):
-        self.screen = utils.load_slide(self.win, slide_image)
-        self.screen.draw()
-        self.win.flip()
-        utils.get_keys(keyList=[wait_key])
-        self.win.flip()
-
-
-class Task_Eyetracker(Task):
-    def __init__(self, eye_tracker=None, target_size=7, **kwargs):
-        super().__init__(**kwargs)
-
-        self.eye_tracker = eye_tracker
-
-        self.mon_size = self.win.monitor.getSizePix()
-        self.SCN_W, self.SCN_H = self.mon_size
-        self.monitor_width = self.win.monitor.getWidth()
-        self.pixpercm = self.mon_size[0] / self.monitor_width
-        self.subj_screendist_cm = self.win.monitor.getDistance()
-
-        self.target_size_pix = self.deg_2_pix(target_size)
-
-        # prepare the pursuit target, the clock and the movement parameters
-        self.win.color = [0, 0, 0]
-        self.win.flip()
-        self.target = visual.GratingStim(
-            self.win, tex=None, mask="circle", size=self.target_size_pix
-        )
-
-    def pos_psych2pix(self, locs: list):
-        """compute location x, y from 0 centered psychopy to top-left centered pixels"""
-        x = int(locs[0] + self.win.size[0] / 2.0)
-        y = int(self.win.size[1] / 2.0 - locs[1])
-        return [x, y]
-
-    def send_target_loc(
-            self, loc: list, target_name="target", to_marker=True, no_interpolation=0
-    ):
-        """send target loc(ation) 0 centered to eyetracker after converting to top-left centered pixels.
-        no_interpolation: 0 or 1
-            0 interpolates, 1 doesn't"""
-        loc = self.pos_psych2pix(loc)
-        self.sendMessage(
-            f"!V TARGET_POS {target_name} {loc[0]}, {loc[1]} 1 {no_interpolation}",
-            to_marker,
-        )  # 1 0  eyetracker code x, y, draw (1 yes), interpolation (0 == yes)
-
-    def deg_2_pix(self, deg):
-        return deg2pix(deg, self.subj_screendist_cm, self.pixpercm)
-
-    def sendMessage(self, msg, to_marker=True, add_event=False):
-        if self.eye_tracker is not None:
-            self.eye_tracker.tk.sendMessage(msg)
-            if to_marker:
-                self.send_marker(msg, add_event)
-
-    def setOfflineMode(self):
-        if self.eye_tracker is not None:
-            self.eye_tracker.paused = True
-            self.eye_tracker.tk.setOfflineMode()
-
-    def startRecording(self):
-        if self.eye_tracker is not None:
-            self.eye_tracker.tk.startRecording(1, 1, 1, 1)
-            self.eye_tracker.paused = False
-
-    def sendCommand(self, msg):
-        if self.eye_tracker is not None:
-            self.eye_tracker.tk.sendCommand(msg)
-
-    def doDriftCorrect(self, vals):
-        # vals : int, position target in screen
-        if self.eye_tracker is not None:
-            self.eye_tracker.tk.doDriftCorrect(*vals)
-
-    def gaze_contingency(self):
-        # move task
-        pass
-
-
-class EyelinkColor(Enum):
-    '''
-       Color codes accepted by the Eyetracker.
-       The source of these codes is from the comments in the examples
-       provided by SR Research - specifically saccade.py
-       Example script can be found in:
-       C:\Program Files (x86)\SR Research\EyeLink\SampleExperiments\Python\examples\Psychopy_examples
-    '''
-    BLACK = 0
-    BLUE = 1
-    GREEN = 2
-    CYAN = 3
-    RED = 4
-    MAGENTA = 5
-    BROWN = 6
-    LIGHTGRAY = 7
-    DARKGRAY = 8
-    LIGHTBLUE = 9
-    LIGHTGREEN = 10
-    LIGHTCYAN = 11
-    LIGHTRED = 12
-    BRIGHTMAGENTA = 13
-    YELLOW = 14
-    BRIGHTWHITE = 15
-
-
-class Eyelink_HostPC(Task_Eyetracker):
-    '''
-       Class containing methods that interact with the HostPC display.
-       HostPC is the computer that controls the eye tracker camera. Tasks
-       can interact with the display that is attached to this computer. In
-       our set up, the display is an Android tablet running the Eyelink app.
-
-       The commands that are sent to the HostPC are in the form of strings,
-       the list of commands and documentation is available in the
-       "COMMANDS.INI" file inside the HostPC (i.e. NUC) - this can be found
-       via the WebUI or by browsing the filesystem in the NUC and looking in
-       /elcl/exe directory. Additional documentation is in the SR Research
-       forums.
-    '''
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    def draw_cross(self, x: int, y: int, colour: EyelinkColor) -> None:
-        '''Draw a cross at the x,y position on the screen of the specified colour
-           x, y must be in the top-left centered coordinate space
-        '''
-        self.sendCommand('draw_cross %d %d %d' % (x, y, colour.value))
-
-    def draw_box(self, x: int, y: int, width: int, height: int, colour: EyelinkColor, filled: bool = False) -> None:
-        '''
-           Draw a rectangle of size width by height (in pixels) around a point
-           x,y on the screen. x, y must be in the top-left centered coordinate
-           space.
-
-           'draw_box' command for the HostPC takes in the diagonal coordinates and
-           a color integer
-        '''
-        half_box_wid_in_pix = int(width/2)
-        half_box_hei_in_pix = int(height/2)
-
-        box_coords_top_x = x-half_box_wid_in_pix
-        box_coords_top_y = y-half_box_hei_in_pix
-        box_coords_bot_x = x+half_box_wid_in_pix
-        box_coords_bot_y = y+half_box_hei_in_pix
-
-        if not filled:
-            self.sendCommand('draw_box %d %d %d %d %d' % (box_coords_top_x, box_coords_top_y,
-                                                               box_coords_bot_x, box_coords_bot_y,
-                                                               colour.value))
-        else:
-            self.sendCommand('draw_filled_box %d %d %d %d %d' % (box_coords_top_x, box_coords_top_y,
-                                                               box_coords_bot_x, box_coords_bot_y,
-                                                               colour.value))
-            pass
-
-    def update_screen(self, x: float, y: float) -> None:
-        '''
-           Draw a cross and a box on the eyelink tablet.
-           x and y are positions in 0 centered psychopy coordinate space.
-           x and y can be int or float
-           See pos_psych2pix above
-        '''
-
-        # First clear tablet screen
-        self.clear_screen()
-
-        # convert from 0 centered to top-left centered coordinate space
-        xy = self.pos_psych2pix([x, y])
-        top_left_x = xy[0]
-        top_left_y = xy[1]
-
-        # draw cross at top_left centered x,y position
-        self.draw_cross(top_left_x, top_left_y, EyelinkColor.LIGHTGREEN)
-
-        # draw box around cross
-        box_len_in_pix = 100
-        self.draw_box(top_left_x, top_left_y, box_len_in_pix, box_len_in_pix, EyelinkColor.LIGHTRED)
-
-    def clear_screen(self, colour: EyelinkColor = EyelinkColor.BLACK) -> None:
-        '''Clear the HostPC screen and leave a Black background by default'''
-        self.sendCommand('clear_screen %d' % colour.value)
-
-    def _render_image(self, path_to_image: Union[str, os.PathLike],
-                     crop_x: int, crop_y: int, crop_width: int, crop_height: int,
-                     host_x: int, host_y: int,
-                     drawing_options: any = pylink.BX_MAXCONTRAST) -> None:
-        '''
-           Employs the imageBackdrop method of eyetracker to display an image on
-           HostPC screen. Documentation available as comments in (and code adapted
-           from) picture.py example provided in SR Research software located at:
-           C:\Program Files (x86)\SR Research\EyeLink\SampleExperiments\Python\examples\Psychopy_examples
-
-           This method:
-           ** DOES NOT SUPPORT IMAGE RESIZING **
-           ** ONLY WORKS WHEN EYETRACKER IS IN OFFLINE MODE **
-
-           Therefore this method needs to be executed before eyetracker starts
-           recording
-
-           :param path_to_image: complete file path to the image file
-           :param crop_x/crop_y: x,y coordinate in pixels for cropping the image
-           :param crop_width/crop_height: width and height in pixels  for cropping
-                                          the image
-           :param host_x/host_y: x,y position in pixels on the HostPC screen where
-                                 the image needs to be rendered
-           :param drawing_options: drawing options from SR Research's pylink package
-
-           Example usage:
-           If you want to render a 1920 by 1080 image on a 1920x1080 screen,
-           crop_x/crop_y would be 0,0, since you do not want to crop anything
-           crop_width/crop_height would be 1920/1080 since you want the full image
-           host_x/host_y would be 0,0 since you want the full image displayed on
-           screen
-
-           For Bamboo Passage the image size is 1536 by 864 which we want to render
-           on a 1920x1080 screen. We do not want to crop the image, but we want to
-           center the image on screen. Therefore the params would be:
-           crop_x/crop_y = 0,0
-           crop_width/crop_height = 1536, 864
-           host_x/host_y = 192, 108 {i.e. (self.SCN_W-1536)/2 & (self.SCN_H-864)/2}
-
-           This method ** DOES NOT SUPPORT ** image resizing. For resizing options
-           follow the el_tracker.bitmapBackdrop() method provided in picture.py
-           example script.
-
-           This method starts with an '_'(underscore) because it is not meant to be
-           called directly, and must be wrapped in a wrapper called 'render_image'
-           within the task script for the image to be rendered when called within
-           server_stm script.
-        '''
-        self.clear_screen()
-
-        if self.eye_tracker is not None:
-            self.eye_tracker.tk.imageBackdrop(path_to_image,
-                                           crop_x, crop_y, crop_width, crop_height,
-                                           host_x, host_y,
-                                           drawing_options)
-
-
-class Task_ShowProgressBar(Task):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.screen = None
-
-    def run(self, **kwargs) -> None:
-        self.screen = utils.load_slide(self.win, kwargs['slide_image'])
-        self.show_text(screen=self.screen, msg="Task", audio=None, wait_time=kwargs['duration'], waitKeys=kwargs['wait_keys'])
-        self.present_complete()
-
-
-class Introduction_Task(Task):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    def run(self, **kwargs):
-        self.present_instructions(prompt=False)
-
-
-if __name__ == "__main__":
-
-    instruction_file = op.join(neurobooth_os.__path__[0], "tasks", "assets", "test.mp4")
-    if not op.isfile(instruction_file):
-        raise IOError(f'Required instruction file {instruction_file} does not exist.')
-
-    task = Task_countdown(
-        instruction_file=instruction_file
-    )
-
-    task.run(prompt=True, duration=3)
