+++ neurobooth_os/tasks/task_introduction.py
@@ -1,0 +1,9 @@
+from neurobooth_os.tasks import Task
+
+
+class Introduction_Task(Task):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def run(self, **kwargs):
+        self.present_instructions(prompt=False)
